requirements.txt
networkx==3.4.2
numpy==2.2.6
packaging==25.0
pillow==11.2.1
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.3
svgwrite==1.4.3
threadpoolctl==3.6.0
tifffile==2025.5.10
opencv-python==4.11.0.86
pytesseract==0.3.13
easyocr==1.7.1
torch==2.3.0
torchvision==0.18.0
openai==1.86.0
python-dotenv==1.1.0
